refactor(indexer): split merkle, git, and content into submodules
lib.rs folded Merkle hashing, git integration, and symbol-content extraction
in with the indexing pipeline. Move each into merkle.rs / git.rs / content.rs;
lib.rs re-exports what index_directory and external callers need
(git_recently_changed_files stays pub). SymbolDiff fields are pub(crate) so
the pipeline can read them across the module boundary. The pub bench_support
module stays in lib.rs (its cartog_indexer::bench_support path is unchanged).

Pure code movement, no API change (36 tests pass, benches still build).
lib.rs drops from 2358 to 1978 lines.
