+++ crates/cartog-indexer/src/lib.rs
@@ -709,348 +709,14 @@ fn dedup_symbol_ids(symbols: &mut [Symbol], edges: &mut [cartog_core::Edge]) {
     );
 }
 
-// ── Merkle-tree hashing ──
+mod content;
+mod git;
+mod merkle;
 
-/// Compute content_hash and subtree_hash for all symbols in an extraction.
-///
-/// - content_hash = sha256(kind + name + signature + body_source)
-/// - subtree_hash = sha256(content_hash + sorted(children_subtree_hashes))
-///
-/// Modifies symbols in-place.
-fn compute_merkle_hashes(symbols: &mut [Symbol], source: &str) {
-    use std::collections::HashMap;
-
-    // Compute content_hash for each symbol
-    for sym in symbols.iter_mut() {
-        let body = source
-            .get(sym.start_byte as usize..sym.end_byte as usize)
-            .unwrap_or("");
-        let mut hasher = Sha256::new();
-        hasher.update(sym.kind.as_str().as_bytes());
-        hasher.update(b":");
-        hasher.update(sym.name.as_bytes());
-        hasher.update(b":");
-        if let Some(ref sig) = sym.signature {
-            hasher.update(sig.as_bytes());
-        }
-        hasher.update(b":");
-        hasher.update(body.as_bytes());
-        sym.content_hash = Some(format!("{:x}", hasher.finalize()));
-    }
-
-    // Build parent→children map by index
-    let id_to_idx: HashMap<&str, usize> = symbols
-        .iter()
-        .enumerate()
-        .map(|(i, s)| (s.id.as_str(), i))
-        .collect();
-
-    let mut children: HashMap<usize, Vec<usize>> = HashMap::new();
-    let mut roots: Vec<usize> = Vec::new();
-
-    for (i, sym) in symbols.iter().enumerate() {
-        if let Some(ref pid) = sym.parent_id {
-            if let Some(&parent_idx) = id_to_idx.get(pid.as_str()) {
-                children.entry(parent_idx).or_default().push(i);
-            } else {
-                roots.push(i);
-            }
-        } else {
-            roots.push(i);
-        }
-    }
-
-    // Post-order traversal to compute subtree_hash bottom-up
-    let mut subtree_hashes: Vec<String> = vec![String::new(); symbols.len()];
-    let mut stack: Vec<(usize, bool)> = roots.iter().rev().map(|&i| (i, false)).collect();
-
-    while let Some((idx, visited)) = stack.pop() {
-        if visited {
-            // All children processed — compute subtree hash
-            let mut hasher = Sha256::new();
-            hasher.update(
-                symbols[idx]
-                    .content_hash
-                    .as_deref()
-                    .unwrap_or("")
-                    .as_bytes(),
-            );
-            if let Some(kids) = children.get(&idx) {
-                let mut kid_hashes: Vec<&str> =
-                    kids.iter().map(|&k| subtree_hashes[k].as_str()).collect();
-                kid_hashes.sort();
-                for h in kid_hashes {
-                    hasher.update(h.as_bytes());
-                }
-            }
-            subtree_hashes[idx] = format!("{:x}", hasher.finalize());
-        } else {
-            stack.push((idx, true));
-            if let Some(kids) = children.get(&idx) {
-                for &kid in kids.iter().rev() {
-                    stack.push((kid, false));
-                }
-            }
-        }
-    }
-
-    // Store subtree_hash in symbols
-    for (i, sym) in symbols.iter_mut().enumerate() {
-        sym.subtree_hash = Some(std::mem::take(&mut subtree_hashes[i]));
-    }
-}
-
-/// Result of diffing new symbols against stored hashes.
-#[derive(Debug, Default)]
-struct SymbolDiff {
-    added: Vec<usize>,            // indices into new symbols
-    removed: Vec<String>,         // IDs to delete from DB
-    modified: Vec<usize>,         // indices into new symbols (content changed)
-    children_changed: Vec<usize>, // indices: own content same, child subtree differs
-    unchanged: usize,             // count of fully unchanged symbols
-}
-
-/// Compare newly extracted symbols against stored hashes for a file.
-fn merkle_diff(
-    new_symbols: &[Symbol],
-    old_hashes: &[(String, Option<String>, Option<String>)],
-) -> SymbolDiff {
-    use std::collections::{HashMap, HashSet};
-
-    let mut diff = SymbolDiff::default();
-
-    let old_map: HashMap<&str, (&Option<String>, &Option<String>)> = old_hashes
-        .iter()
-        .map(|(id, ch, sh)| (id.as_str(), (ch, sh)))
-        .collect();
-
-    let new_ids: HashSet<&str> = new_symbols.iter().map(|s| s.id.as_str()).collect();
-
-    for (i, sym) in new_symbols.iter().enumerate() {
-        if let Some(&(old_ch, old_sh)) = old_map.get(sym.id.as_str()) {
-            // Symbol exists in both old and new
-            if sym.subtree_hash.as_ref() == old_sh.as_ref()
-                && sym.content_hash.as_ref() == old_ch.as_ref()
-            {
-                diff.unchanged += 1;
-            } else if sym.content_hash.as_ref() != old_ch.as_ref() {
-                diff.modified.push(i);
-            } else {
-                // content same, subtree different — a child was added/modified/removed
-                diff.children_changed.push(i);
-            }
-        } else {
-            diff.added.push(i);
-        }
-    }
-
-    for (old_id, _, _) in old_hashes {
-        if !new_ids.contains(old_id.as_str()) {
-            diff.removed.push(old_id.clone());
-        }
-    }
-
-    diff
-}
-
-/// Get list of files changed since the last indexed commit.
-///
-/// Returns `None` (triggering hash fallback) when:
-/// - `last_commit` is `None` (first index)
-/// - Not inside a git repository
-/// - The stored commit no longer exists (after rebase/reset)
-fn git_changed_files(
-    root: &Path,
-    last_commit: Option<&str>,
-) -> Option<std::collections::HashSet<String>> {
-    let last_commit = last_commit?;
-
-    // Verify the stored commit still exists in history
-    let verify = git_cmd(root, &["cat-file", "-t", last_commit])?;
-    if !verify.status.success() {
-        return None;
-    }
-
-    // Get files changed between last indexed commit and HEAD
-    let diff_output = git_cmd(root, &["diff", "--name-only", last_commit, "HEAD"])?;
-    if !diff_output.status.success() {
-        return None;
-    }
-
-    let mut changed: std::collections::HashSet<String> =
-        parse_git_lines(&diff_output.stdout).collect();
-
-    // Also include untracked files (new files not yet committed)
-    if let Some(out) = git_cmd(root, &["ls-files", "--others", "--exclude-standard"]) {
-        if out.status.success() {
-            changed.extend(parse_git_lines(&out.stdout));
-        }
-    }
-
-    // Also include unstaged/staged changes in the working tree
-    if let Some(out) = git_cmd(root, &["diff", "--name-only"]) {
-        if out.status.success() {
-            changed.extend(parse_git_lines(&out.stdout));
-        }
-    }
-
-    if let Some(out) = git_cmd(root, &["diff", "--name-only", "--cached"]) {
-        if out.status.success() {
-            changed.extend(parse_git_lines(&out.stdout));
-        }
-    }
-
-    Some(changed)
-}
-
-/// Get the current HEAD commit hash.
-fn git_head_commit(root: &Path) -> Option<String> {
-    let output = git_cmd(root, &["rev-parse", "HEAD"])?;
-    if output.status.success() {
-        Some(String::from_utf8(output.stdout).ok()?.trim().to_string())
-    } else {
-        None
-    }
-}
-
-/// Get files changed in the last N commits + working tree changes (staged, unstaged, untracked).
-///
-/// Returns a sorted, deduplicated list of file paths relative to `root`.
-/// Returns `Err` if not inside a git repository.
-pub fn git_recently_changed_files(root: &Path, commits: u32) -> Result<Vec<String>> {
-    use std::collections::BTreeSet;
-    let mut changed = BTreeSet::new();
-
-    // Files changed in last N commits
-    let output = git_cmd(
-        root,
-        &[
-            "log",
-            "--name-only",
-            "--pretty=format:",
-            &format!("-{commits}"),
-        ],
-    )
-    .context("Failed to run git — are you in a git repository?")?;
-    if output.status.success() {
-        changed.extend(parse_git_lines(&output.stdout));
-    }
-
-    // Working tree changes (unstaged + staged + untracked)
-    for args in [
-        &["diff", "--name-only"][..],
-        &["diff", "--name-only", "--cached"][..],
-        &["ls-files", "--others", "--exclude-standard"][..],
-    ] {
-        if let Some(out) = git_cmd(root, args) {
-            if out.status.success() {
-                changed.extend(parse_git_lines(&out.stdout));
-            }
-        }
-    }
-
-    Ok(changed.into_iter().collect())
-}
-
-/// Run a git command with stdin suppressed to prevent interactive prompts.
-fn git_cmd(root: &Path, args: &[&str]) -> Option<std::process::Output> {
-    std::process::Command::new("git")
-        .args(args)
-        .current_dir(root)
-        .stdin(std::process::Stdio::null())
-        .output()
-        .ok()
-}
-
-/// Parse lines from git command output, filtering empty lines.
-fn parse_git_lines(stdout: &[u8]) -> impl Iterator<Item = String> + '_ {
-    String::from_utf8_lossy(stdout)
-        .lines()
-        .filter(|l| !l.is_empty())
-        .map(|l| l.to_string())
-        .collect::<Vec<_>>()
-        .into_iter()
-}
-
-/// Find the largest byte index <= `index` that is a valid UTF-8 char boundary in `s`.
-///
-/// Equivalent to the nightly `str::floor_char_boundary`. Walks back at most 3 bytes.
-fn floor_char_boundary(s: &str, index: usize) -> usize {
-    if index >= s.len() {
-        return s.len();
-    }
-    // UTF-8 continuation bytes have the pattern 10xxxxxx (0x80..0xBF).
-    // Walk backwards until we find a byte that is NOT a continuation byte.
-    let mut i = index;
-    while i > 0 && !s.is_char_boundary(i) {
-        i -= 1;
-    }
-    i
-}
-
-/// Maximum content length (in bytes) stored per symbol for embedding.
-///
-/// BERT models have a 512-token limit. Code averages ~1 token per 2-3 chars,
-/// so 2048 bytes ≈ 680-1024 tokens, truncated to 512 by the model.
-/// This captures the symbol signature + leading body while halving inference time.
-const MAX_CONTENT_BYTES: usize = 2048;
-
-/// Minimum content length (in bytes) to bother embedding.
-///
-/// Symbols shorter than this (e.g. `import os`, `x = 1`) add noise without value.
-const MIN_CONTENT_BYTES: usize = 50;
-
-/// Extract the raw source code for a symbol and build a metadata header.
-///
-/// Returns `(content, header)` where `header` is a brief preamble for embedding context.
-/// Returns `None` if: byte offsets are invalid, content is empty/too short,
-/// or the symbol is an import (not useful for semantic search).
-fn extract_symbol_content(source: &str, sym: &cartog_core::Symbol) -> Option<(String, String)> {
-    // Skip imports — they don't contain searchable logic.
-    if sym.kind == cartog_core::SymbolKind::Import {
-        return None;
-    }
-
-    let start = sym.start_byte as usize;
-    let end = sym.end_byte as usize;
-
-    if start >= end || end > source.len() {
-        return None;
-    }
-
-    // Ensure both boundaries fall on valid UTF-8 char boundaries.
-    // Tree-sitter should produce valid offsets, but truncation at MAX_CONTENT_BYTES
-    // can land mid-character for multi-byte content (e.g. '─' = 3 bytes).
-    let safe_start = if source.is_char_boundary(start) {
-        start
-    } else {
-        // Ceil to next char boundary
-        let mut s = start;
-        while s < source.len() && !source.is_char_boundary(s) {
-            s += 1;
-        }
-        s
-    };
-    let truncated_end = end.min(safe_start + MAX_CONTENT_BYTES);
-    let safe_end = floor_char_boundary(source, truncated_end);
-
-    if safe_start >= safe_end {
-        return None;
-    }
-
-    let raw = &source[safe_start..safe_end];
-    let trimmed = raw.trim();
-    if trimmed.is_empty() || trimmed.len() < MIN_CONTENT_BYTES {
-        return None;
-    }
-
-    let header = format!(
-        "// File: {}\n// Type: {}\n// Name: {}",
-        sym.file_path, sym.kind, sym.name
-    );
-
-    Some((raw.to_string(), header))
-}
+pub(crate) use content::extract_symbol_content;
+pub use git::git_recently_changed_files;
+pub(crate) use git::{git_changed_files, git_head_commit};
+pub(crate) use merkle::{compute_merkle_hashes, merkle_diff};
 
 /// Shared scenario bodies for the indexing benchmarks.
 ///
@@ -1596,52 +1262,6 @@ mod tests {
             3,
             "no-op reindex must not reset state=3"
         );
-    }
-
-    #[test]
-    fn test_floor_char_boundary_ascii() {
-        let s = "hello world";
-        assert_eq!(floor_char_boundary(s, 5), 5);
-        assert_eq!(floor_char_boundary(s, 0), 0);
-        assert_eq!(floor_char_boundary(s, 100), s.len());
-    }
-
-    #[test]
-    fn test_floor_char_boundary_multibyte() {
-        // '─' is U+2500, encoded as 3 bytes: E2 94 80
-        let s = "abc─def";
-        // a=0, b=1, c=2, ─=3..6, d=6, e=7, f=8
-        assert_eq!(floor_char_boundary(s, 3), 3); // start of ─
-        assert_eq!(floor_char_boundary(s, 4), 3); // mid ─ → snap back
-        assert_eq!(floor_char_boundary(s, 5), 3); // mid ─ → snap back
-        assert_eq!(floor_char_boundary(s, 6), 6); // start of 'd'
-    }
-
-    #[test]
-    fn test_extract_symbol_content_truncates_at_char_boundary() {
-        // Build a source string where MAX_CONTENT_BYTES truncation lands mid-char.
-        // Fill with ASCII up to MAX_CONTENT_BYTES-1, then add a 3-byte char.
-        let padding = "x".repeat(MAX_CONTENT_BYTES - 1);
-        let source = format!("{padding}─after");
-
-        let sym = cartog_core::Symbol::new(
-            "test_sym",
-            cartog_core::SymbolKind::Function,
-            "test.rb",
-            1,
-            100,
-            0,
-            source.len() as u32,
-            None,
-        );
-
-        // This should NOT panic despite truncation landing inside '─'
-        let result = extract_symbol_content(&source, &sym);
-        assert!(result.is_some());
-        let (content, _header) = result.unwrap();
-        // Content should be truncated before the '─' (snapped to char boundary)
-        assert_eq!(content.len(), MAX_CONTENT_BYTES - 1);
-        assert!(content.is_char_boundary(content.len()));
     }
 
     // ── Dedup tests ──
